requirements.txt
beautifulsoup4==4.12.2
pandas==1.2.3
python-dotenv==1.0.0
requests==2.22.0